refactor some code and try integrating with Unity
