+++ livekit-ffi/protocol/audio_frame.proto
@@ -286,7 +286,11 @@ message AudioSourceOptions {
 }
 
 enum AudioSourceType {
+  // Push-based audio source - manually capture frames via CaptureAudioFrameRequest
   AUDIO_SOURCE_NATIVE = 0;
+  // Platform ADM-based audio source - captures from microphone automatically
+  // Requires PlatformAudio to be created first to enable ADM recording
+  AUDIO_SOURCE_PLATFORM = 1;
 }
 
 message AudioSourceInfo {
