+++ livekit/src/prelude.rs
@@ -36,7 +36,7 @@ pub use crate::{
 
 // Platform audio device management (native platforms only)
 #[cfg(not(target_arch = "wasm32"))]
-pub use crate::audio::{
+pub use crate::platform_audio::{
     AudioError, AudioProcessingOptions, AudioProcessingType, AudioResult, PlatformAudio,
     RtcAudioSource,
 };
